+++ src/features/account/components/Nav.tsx
@@ -1,10 +1,10 @@
 import { Link } from 'react-router-dom';
 
-interface NavProps {
-  handleLogout: () => void;
-}
+import { useAuthContext } from '@contexts/AuthContext';
 
-export default function Nav({ handleLogout }: NavProps) {
+export default function Nav() {
+  const { signOut } = useAuthContext();
+
   return (
     <nav className="max-h-fit basis-1/4 rounded-md bg-white p-5 md:p-7.5 lg:p-10">
       <ul className="flex flex-col gap-y-7.5 text-sm tracking-xwide uppercase">
@@ -29,7 +29,7 @@ export default function Nav({ handleLogout }: NavProps) {
           </Link>
         </li>
         <li>
-          <Link to="" onClick={handleLogout} className="text-shark transition-colors duration-300 ease-in-out hover:text-boulder">
+          <Link to="" onClick={signOut} className="text-shark transition-colors duration-300 ease-in-out hover:text-boulder">
             Log out
           </Link>
         </li>